Making new shadow block class for #151
