+++ src/controller/toolbox_controller.js
@@ -31,14 +31,16 @@ goog.require('ToolboxEditorView');
  *
  * @authors sagev (Sage Vouse), celinechoo (Celine Choo), evd2014 (Emma Dauterman)
  */
-class ToolboxController {
+class ToolboxController extends ShadowController {
+  /**
+   * Manages toolbox editor.
+   * @param {!ProjectController} projectController ProjectController object which
+   *     will make changes to current project upon user's changes in editor.
+   * @param {!Blockly.Workspace} hiddenWorkspace Hidden blockly workspace used
+   *     to create shadow blocks.
+   */
   constructor(projectController, hiddenWorkspace) {
-    /**
-     * ProjectController that will edit toolboxes when edited in the toolbox
-     *     editor.
-     * @type {!ProjectController}
-     */
-    this.projectController = projectController;
+    super(projectController, hiddenWorkspace);
 
     // Creates first toolbox to add to project.
     const toolbox = this.projectController.createToolbox('MyFirstToolbox');
@@ -63,11 +65,8 @@ class ToolboxController {
      */
     this.keyEventsEnabled = true;
 
-    /**
-     * Hidden Blockly workspace used to generate XML for shadow blocks.
-     * @type {!Blockly.Workspace}
-     */
-    this.hiddenWorkspace = hiddenWorkspace;
+    // Sets current resource for shadow block class.
+    this.setResource(this.view.toolbox);
 
     // Adds current toolbox to model.
     this.projectController.addToolbox(this.view.toolbox);
@@ -549,107 +548,6 @@ class ToolboxController {
     }
   }
 
-  /*
-   * Makes the currently selected block a user-generated shadow block. These
-   * blocks are not made into real shadow blocks, but recorded in the model
-   * and visually marked as shadow blocks, allowing the user to move and edit
-   * them (which would be impossible with actual shadow blocks). Updates the
-   * preview when done.
-   */
-  setSelectedAsShadowBlock() {
-    // From wfactory_controller.js:addShadow()
-    // No block selected to make a shadow block.
-    if (!this.view.selectedBlock) {
-      return;
-    }
-    FactoryUtils.markShadowBlock(this.view.selectedBlock);
-    this.view.toolbox.addShadowBlock(this.view.selectedBlock.id);
-
-    // Apply shadow block to the children as well.
-    for (let block of this.view.selectedBlock.getDescendants()) {
-      FactoryUtils.markShadowBlock(block);
-      this.view.toolbox.addShadowBlock(block.id);
-    }
-
-    this.view.showAndEnableShadow(false,
-        FactoryUtils.isValidShadowBlock(this.view.selectedBlock, true));
-    this.checkShadowStatus();
-    // Save and update the preview.
-    this.saveStateFromWorkspace();
-    this.updatePreview();
-  }
-
-  /**
-   * Makes user-generated shadow block back to a normal block again. Removes
-   * block from list of shadow blocks and then reloads workspace. Updates the
-   * preview when done.
-   */
-  unsetSelectedAsShadowBlock() {
-    // From wfactory_controller.js
-    if (!this.view.selectedBlock) {
-      return;
-    }
-    FactoryUtils.unmarkShadowBlock(this.view.selectedBlock);
-    this.view.toolbox.removeShadowBlock(this.view.selectedBlock.id);
-    this.checkShadowStatus();
-    this.view.showAndEnableShadow(true,
-        FactoryUtils.isValidShadowBlock(this.view.selectedBlock, true));
-
-    // Save and update the preview.
-    this.saveStateFromWorkspace();
-    this.updatePreview();
-  }
-
-  /**
-   * Checks currently selected block if it is breaking any shadow block rules.
-   * Sets warning text to user if it breaks a rule, and removes warning
-   * text if not.
-   * Shadow blocks must be nested within another block and cannot have any
-   * non-shadow blocks as children.
-   * @recursive Checks connected blocks for their shadow block status.
-   */
-  checkShadowStatus() {
-    const selected = this.view.selectedBlock;
-    if (!selected) {
-      // Return if no block is selected.
-      return;
-    }
-
-    // Check if shadow block.
-    const isShadow = FactoryUtils.isUserGenShadowBlock(selected.id, this.view.toolbox) ||
-        $(selected.svgGroup_).hasClass('shadowBlock');
-
-    // Check if valid shadow block position.
-    const isValid = FactoryUtils.isValidShadowBlock(selected, isShadow);
-
-    // Check if block has variables (variable blocks cannot be shadow blocks).
-    const hasVar = FactoryUtils.hasVariableField(selected);
-    this.view.enableShadowButtons(isShadow, isValid);
-
-    // Checks various states of the selected block to make the warning text
-    // more helpful/descriptive. Removes warning text if no rules are broken.
-    if (isShadow && !isValid) {
-      selected.setWarningText('Shadow blocks must be nested inside\n'
-          + 'other shadow blocks or regular blocks.');
-    } else if (isShadow && hasVar) {
-      selected.setWarningText('Shadow blocks must be nested inside other'
-          + ' blocks.');
-    } else if (!isShadow && selected.getSurroundParent() &&
-        $(selected.getSurroundParent().svgGroup_).hasClass('shadowBlock')) {
-      selected.setWarningText('Regular blocks cannot be children\nof shadow '
-          + 'blocks.');
-    } else {
-      selected.setWarningText(null);
-    }
-
-    if (selected.getNextBlock()) {
-      // Recursively check next block in connection for its status.
-      this.view.selectedBlock = selected.getNextBlock();
-      this.checkShadowStatus();
-      this.view.selectedBlock = selected;
-    }
-  }
-
   /**
    * Convert actual shadow blocks added from the toolbox to user-generated shadow
    * blocks.
@@ -657,8 +555,7 @@ class ToolboxController {
    * @private
    */
   createFakeShadowBlocks_(blockId) {
-    // Converts actual shadow blocks to shadow-looking blocks in editor.
-    this.convertShadowBlocks();
+    super.createFakeShadowBlocks_(blockId);
 
     // Let the user create a Variables or Functions category if they use
     // blocks from either category.
@@ -710,123 +607,12 @@ Do you want to add a ${categoryName} category to your custom toolbox?`;
   }
 
   /**
-   * Determines if a block breaks shadow block placement rules. Breaks rules
-   * if (1) a shadow block no longer has a valid parent, or (2) a normal block
-   * is inside of a shadow block.
-   * @param {!Blockly.Block} block Blockly block that will be checked.
-   * @return {boolean} Whether block is placed in valid location as a shadow block.
-   * @private
-   */
-  isInvalidBlockPlacement_(block) {
-    const isShadow = FactoryUtils.isUserGenShadowBlock(block.id, this.view.toolbox);
-    return ((isShadow && !block.getSurroundParent()) ||
-        (!isShadow && block.getSurroundParent() &&
-         FactoryUtils.isUserGenShadowBlock(block.getSurroundParent().id, this.view.toolbox)));
-  }
-
-  /**
    * Saves the contents of the toolbox editor workspace to the corresponding
    * ListElement. Updates the XML in the model.
    */
   saveStateFromWorkspace() {
     this.view.toolbox.getSelected().saveFromWorkspace(this.view.editorWorkspace);
     this.view.toolbox.setXml(this.generateToolboxXml());
-  }
-
-  /**
-   * Loads the given XML to the hidden Blockly.Workspace and sets any user-generated
-   * shadow blocks to be actual shadow blocks in the hidden Blockly.Workspace.
-   *
-   * @param {!Element} xml XML to be loaded to the hidden workspace.
-   * @private
-   */
-  loadToHiddenWorkspace_(xml) {
-    // REFACTOR: Moved in from wfactory_generator.js
-    this.hiddenWorkspace.clear();
-    Blockly.Xml.domToWorkspace(xml, this.hiddenWorkspace);
-    this.setShadowBlocksInHiddenWorkspace_();
-  }
-
-  /**
-   * Encodes blocks in the hidden workspace in a XML DOM element. Very
-   * similar to workspaceToDom, but doesn't capture IDs. Uses the top-level
-   * blocks loaded in hiddenWorkspace.
-   * @param {!Element} xmlDom Tree of XML elements to be appended to.
-   * @private
-   */
-  appendHiddenWorkspaceToDom_(xmlDom) {
-    // REFACTOR: Moved in from wfactory_generator.js
-    const blocks = this.hiddenWorkspace.getTopBlocks();
-    blocks.forEach((block) => {
-      let blockChild = Blockly.Xml.blockToDom(block, /* opt_noId */ true);
-      xmlDom.appendChild(blockChild);
-    });
-  }
-
-  /**
-   * Sets the user-generated shadow blocks loaded into hiddenWorkspace to be
-   * actual shadow blocks. This is done so that blockToDom records them as
-   * shadow blocks instead of regular blocks.
-   * @private
-   */
-  setShadowBlocksInHiddenWorkspace_() {
-    // REFACTOR: Moved in from wfactory_generator.js
-    const blocks = this.hiddenWorkspace.getAllBlocks();
-    blocks.forEach((block) => {
-      if (this.view.toolbox.isShadowBlock(block.id)) {
-        block.setShadow(true);
-      }
-    });
-  }
-
-  /**
-   * Sets a block and all of its children to be user-generated shadow blocks,
-   * both in the model and view.
-   * @param {!Blockly.Block} block The block to be converted to a user-generated
-   *     shadow block.
-   * @private
-   */
-  addShadowForBlockAndChildren_(block) {
-    // From wfactory_controller.js:addShadowForBlockAndChildren_(block)
-    // Convert to shadow block.
-    FactoryUtils.markShadowBlock(block);
-    this.view.toolbox.addShadowBlock(block.id);
-
-    if (FactoryUtils.hasVariableField(block)) {
-      block.setWarningText('Cannot make variable blocks shadow blocks.');
-    }
-
-    // Convert all children to shadow blocks recursively.
-    const children = block.getChildren();
-    children.forEach((child) => {
-      this.addShadowForBlockAndChildren_(child);
-    });
-  }
-
-  /**
-   * Call when importing XML containing real shadow blocks. This function turns
-   * all real shadow blocks loaded in the workspace into user-generated shadow
-   * blocks, meaning they are marked as shadow blocks by the model and appear as
-   * shadow blocks in the view but are still editable and movable.
-   */
-  convertShadowBlocks() {
-    // REFACTORED: Moved in from wfactory_controller.js
-    const blocks = this.view.editorWorkspace.getAllBlocks();
-    blocks.forEach((block) => {
-      if (block.isShadow()) {
-        block.setShadow(false);
-        // Delete the shadow DOM attached to the block so that the shadow block
-        // does not respawn. Dependent on implementation details.
-        const parentConnection = block.outputConnection ?
-            block.outputConnection.targetConnection :
-            block.previousConnection.targetConnection;
-        if (parentConnection) {
-          parentConnection.setShadowDom(null);
-        }
-        this.view.toolbox.addShadowBlock(block.id);
-        FactoryUtils.markShadowBlock(block);
-      }
-    });
   }
 
   /**
@@ -1203,23 +989,5 @@ Do you want to add a ${categoryName} category to your custom toolbox?`;
     this.view.setBorderColor(this.view.toolbox.getSelectedId(), color);
     this.view.toolbox.setXml(this.generateToolboxXml());
     this.updatePreview();
-  }
-
-  /**
-   * Given a set of blocks currently loaded, returns all blocks in the workspace
-   * that are user generated shadow blocks.
-   * @param {Array.<!Blockly.Block>} blocks Array of blocks currently loaded.
-   * @return {Array.<!Blockly.Block>} Array of user-generated shadow blocks currently
-   *     loaded.
-   */
-  getShadowBlocksInWorkspace(blocks) {
-    // From wfactory_model.js:getShadowBlocksInWorkspace()
-    let shadowsInWorkspace = [];
-    blocks.forEach((block) => {
-      if (this.view.toolbox.isShadowBlock(block.id)) {
-        shadowsInWorkspace.push(block);
-      }
-    });
-    return shadowsInWorkspace;
   }
 }
