+++ src/factory_utils.js
@@ -1313,89 +1313,6 @@ FactoryUtils.ifCheckedEnable = function(enabled, idArray) {
 };
 
 /**
- * Returns array of shadow blocks from a list of blocks.
- * @param {!Array.<!Blockly.Block>} blockList List of blocks.
- * @return {!Array.<!Blockly.Block>} List of shadow blocks from given list.
- */
-FactoryUtils.getShadowBlocks = function(blockList) {
-  let shadowBlocks = [];
-  for (let block of blockList) {
-    if ($(block.svgGroup_).hasClass('shadowBlock')) {
-      shadowBlocks.push(block);
-    }
-  }
-  return shadowBlocks;
-};
-
-/**
- * Removes visual marking for a shadow block given a rendered block.
- * @param {!Blockly.Block} block The block that should be unmarked as a shadow
- *     block (must be rendered).
- */
-FactoryUtils.unmarkShadowBlock = function(block) {
-  // REFACTOR: Moved in from wfactory_view.js
-  Blockly.utils.removeClass(block.svgGroup_, 'shadowBlock');
-}
-
-/**
- * Given a Blockly.Block, visually marks a block in the view to look like a
- * shadow block.
- * @param {!Blockly.Block} block Blockly block to be marked as a shadow block.
- */
-FactoryUtils.markShadowBlock = function(block) {
-  // REFACTOR: Moved in from wfactory_view.js:markShadowBlock(block)
-  // Add Blockly CSS for user-generated shadow blocks.
-  Blockly.utils.addClass(block.svgGroup_, 'shadowBlock');
-  // If not a valid shadow block, add a warning message.
-  if (!block.getSurroundParent()) {
-    block.setWarningText('Shadow blocks must be nested inside other blocks' +
-        ' be displayed.');
-  }
-  if (FactoryUtils.hasVariableField(block)) {
-    block.setWarningText('Cannot make variable blocks shadow blocks.');
-  }
-};
-
-/**
- * Given a unique block ID, uses the model to determine if a block is a
- * user-generated shadow block.
- * @param {string} blockId The unique ID of the block to examine.
- * @param {!Toolbox|!WorkspaceContents} resource The resource (which has a
- *     this.shadowBlocks field) to check for shadow blocks.
- * @return {boolean} True if the block is a user-generated shadow block, false
- *     otherwise.
- * @throws {Error} Will throw an error if the given resource is not of type
- *     Toolbox or WorkspaceContents.
- */
-FactoryUtils.isUserGenShadowBlock = function(blockId, resource) {
-  if (resource instanceof Toolbox || resource instanceof WorkspaceContents) {
-    return resource.isShadowBlock(blockId);
-  } else {
-    throw new Error('The given resource, ' + resource.name + ', is not a ' +
-        'Toolbox or a WorkspaceContents object, and therefore cannot store ' +
-        'shadow blocks.');
-  }
-};
-
-/**
- * Checks whether the given block is a valid shadow block.
- * @param {!Blockly.Block} block The block to be evaluated for shadow block
- *     validity.
- * @param {boolean} isShadow Whether the given block is a shadow block.
- * @return {boolean} Whether the given block is a valid shadow block that
- *     does not need warning text.
- */
-FactoryUtils.isValidShadowBlock = function(block, isShadow) {
-  // Check if valid shadow block position.
-  const children = block.getChildren();
-  // To be a valid shadow block candidate, the block must (1) have a parent,
-  // and (2) have only shadow blocks as its children.
-  const isValid = block.getSurroundParent() != null &&
-      (isShadow || FactoryUtils.getShadowBlocks(children).length == children.length);
-  return isValid;
-};
-
-/**
  * Sets a warning on blocks that are not defined within a given project.
  * @param {!Array.<Blockly.Block>} blocks A list of blocks to check if they are
  *     defined.
